Split semantic module for CI loc gate
