+++ crates/ctx-cli/tests/support/runner.rs
@@ -58,9 +58,7 @@ pub(crate) fn copied_ctx_binary(temp: &TempDir) -> PathBuf {
     } else {
         "ctx-test-copy"
     });
-    if fs::hard_link(&source, &target).is_err() {
-        fs::copy(&source, &target).unwrap();
-    }
+    fs::copy(&source, &target).unwrap();
     #[cfg(unix)]
     {
         use std::os::unix::fs::PermissionsExt;
