+++ scripts/check-loc-exceptions.tsv
@@ -8,5 +8,10 @@ crates/ctx-history-capture/src/tests/codex.rs	1844	test	Codex parser regression 
 crates/ctx-history-capture/src/tests/native_sqlite.rs	1656	test	native SQLite provider regression matrix shares setup across formats; split after fixture helpers are shared	2026-10-01
 crates/ctx-history-capture/src/provider/native.rs	1146	source	native provider discovery dispatcher remains centralized until provider registry extraction lands	2026-10-01
 crates/ctx-history-capture/src/provider/importer.rs	1127	source	native provider import orchestration remains centralized until importer phase extraction lands	2026-10-01
-crates/ctx-cli/tests/native_providers.rs	1679	test	native provider public CLI scenario matrix; fixtures are already modular and splitting would scatter end-to-end coverage	2026-10-01
+crates/ctx-cli/tests/native_providers.rs	1704	test	native provider public CLI scenario matrix; fixtures are already modular and splitting would scatter end-to-end coverage	2026-10-01
 crates/ctx-history-capture/src/provider/codex/session.rs	1008	source	Codex session parser recently grew oversized-record semantics; keep bounded while follow-up parser split is scoped	2026-10-01
+crates/ctx-cli/src/commands/import/native.rs	1110	source	native provider dispatcher plus manifest helpers inherited over source cap; immediate split would couple CI gate to importer redesign	2026-10-01
+crates/ctx-history-store/src/search/projections.rs	2021	source	search projection eligibility and embedding document helpers are temporarily over cap after semantic projection integration; split with next search-store cleanup	2026-10-01
+crates/ctx-cli/src/main.rs	1150	source	public CLI argument and dispatch surface is temporarily over cap after daemon/status/semantic flags; split command args with next CLI modularization	2026-10-01
+crates/ctx-cli/src/semantic/daemon.rs	1446	source	daemon lifecycle, query service, and semantic scheduler are integrated for prerelease; split after daemon service boundaries settle	2026-10-01
+crates/ctx-history-store/src/schema/migrations.rs	1043	source	schema migration list is append-only and slightly over cap; split migration helpers with next storage cleanup	2026-10-01
